+++ include/alpaka/core/Common.hpp
@@ -83,77 +83,3 @@
 #else
     #define ALPAKA_FN_INLINE inline
 #endif
-
-namespace alpaka
-{
-    namespace core
-    {
-        namespace detail
-        {
-            //#############################################################################
-            //!
-            //#############################################################################
-            template<
-                typename TArg,
-                typename TSfinae = void>
-            struct AssertValueUnsigned;
-            //#############################################################################
-            //!
-            //#############################################################################
-            template<
-                typename TArg>
-            struct AssertValueUnsigned<
-                TArg,
-                typename std::enable_if<!std::is_unsigned<TArg>::value>::type>
-            {
-                ALPAKA_NO_HOST_ACC_WARNING
-                ALPAKA_FN_HOST_ACC static auto assertValueUnsigned(
-                    TArg const & arg)
-                -> void
-                {
-#ifdef NDEBUG
-                    boost::ignore_unused(arg);
-#else
-                    assert(arg >= 0);
-#endif
-                }
-            };
-            //#############################################################################
-            //!
-            //#############################################################################
-            template<
-                typename TArg>
-            struct AssertValueUnsigned<
-                TArg,
-                typename std::enable_if<std::is_unsigned<TArg>::value>::type>
-            {
-                ALPAKA_NO_HOST_ACC_WARNING
-                ALPAKA_FN_HOST_ACC static auto assertValueUnsigned(
-                    TArg const & arg)
-                -> void
-                {
-#ifndef __CUDA_ARCH__
-                    boost::ignore_unused(arg);
-#endif
-                    // Nothing to do for unsigned types.
-                }
-            };
-        }
-        //-----------------------------------------------------------------------------
-        //! This method checks integral values if they are greater or equal zero.
-        //! The implementation prevents warnings for checking this for unsigned types.
-        //-----------------------------------------------------------------------------
-        template<
-            typename TArg>
-        ALPAKA_NO_HOST_ACC_WARNING
-        ALPAKA_FN_HOST_ACC auto assertValueUnsigned(
-            TArg const & arg)
-        -> void
-        {
-            detail::AssertValueUnsigned<
-                TArg>
-            ::assertValueUnsigned(
-                arg);
-        }
-    }
-}
