+++ include/alpaka/mem/alloc/AllocCpuNew.hpp
@@ -25,6 +25,8 @@
 
 #include <alpaka/core/Common.hpp>       // ALPAKA_FN_HOST
 
+#include <boost/core/ignore_unused.hpp> // boost::ignore_unused
+
 namespace alpaka
 {
     namespace mem
